Moved cmd app files to different folder; updated readme;
