+++ src/main/java/org/openscience/cdk/tools/efgf/app/ExtractFunctionalGroupsTask.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package de.ibci.ertlfxgroupsfinder.performancetest;
+package org.openscience.cdk.tools.efgf.app;
 
 import org.openscience.cdk.interfaces.IAtomContainer;
 import org.openscience.cdk.tools.ErtlFunctionalGroupsFinder;
@@ -30,6 +30,7 @@ import java.util.concurrent.Callable;
  * ErtlFunctionalGroupsFinder class.
  *
  * @author Jonas Schaub
+ * @version 1.2
  */
 public class ExtractFunctionalGroupsTask implements Callable<Integer> {
